Rearranged folders and changed routing to almost work
