+++ package.json
@@ -1,0 +1,25 @@
+{
+  "name": "picky-project",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "dependencies": {
+    "mongodb": "^2.2.30",
+    "mongoose": "^4.11.5",
+    "react-router": "^4.1.2"
+  },
+  "devDependencies": {},
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/clener/picky.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/clener/picky/issues"
+  },
+  "homepage": "https://github.com/clener/picky#readme"
+}
